fix(report): move compare logic into core
