+++ src/core/markdown.rs
@@ -1,0 +1,3 @@
+pub(crate) fn escape_markdown_table_cell(value: &str) -> String {
+    value.replace('|', "\\|").replace('\n', " ")
+}
