+++ src/core/mod.rs
@@ -93,6 +93,7 @@ pub mod issues;
 pub mod keychain;
 pub mod lab_routing;
 pub mod lifecycle;
+pub(crate) mod markdown;
 pub mod observation;
 pub mod output;
 pub(crate) mod ownership;
@@ -111,6 +112,8 @@ pub mod proof;
 pub mod publication_artifacts;
 pub mod quality;
 pub mod redaction;
+pub mod report_compare;
+pub(crate) mod report_compare_render;
 pub mod refactor;
 pub mod release;
 pub mod review;
